requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aiohttp==3.8.4
aiosignal==1.3.1; python_version >= '3.7'
alembic==1.11.1
anyio==3.7.0; python_version >= '3.7'
async-timeout==4.0.2; python_version >= '3.6'
asyncpg==0.27.0
attrs==23.1.0; python_version >= '3.7'
bcrypt==4.0.1
certifi==2023.5.7; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==3.1.0; python_version >= '3.7'
click==8.1.3; python_version >= '3.7'
cryptography==41.0.1
dnspython==2.3.0; python_version >= '3.7' and python_version < '4'
ecdsa==0.18.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
email-validator==2.0.0.post2
fastapi[all]==0.99.1
frozenlist==1.3.3; python_version >= '3.7'
greenlet==2.0.2; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0; python_version >= '3.7'
httpcore==0.17.2; python_version >= '3.7'
httptools==0.5.0
httpx==0.24.1
idna==3.4; python_version >= '3.5'
iniconfig==2.0.0; python_version >= '3.7'
itsdangerous==2.1.2
jinja2==3.1.2
mako==1.2.4; python_version >= '3.7'
markupsafe==2.1.3; python_version >= '3.7'
multidict==6.0.4; python_version >= '3.7'
orjson==3.9.1
packaging==23.1; python_version >= '3.7'
passlib[bcrypt]==1.7.4
pluggy==1.2.0; python_version >= '3.7'
psycopg2-binary==2.9.6
pyasn1==0.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pycparser==2.21
pydantic==1.10.10; python_version >= '3.7'
pytest-asyncio==0.21.0
pytest==7.4.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pyyaml==6.0
rsa==4.9; python_version >= '3.6' and python_version < '4'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0; python_version >= '3.7'
sqlalchemy==2.0.17
starlette==0.27.0; python_version >= '3.7'
typing-extensions==4.7.1; python_version >= '3.7'
ujson==5.8.0
uvicorn[standard]==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2; python_version >= '3.7'
gunicorn